requirements.txt
alembic==1.4.1
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
awscli==1.18.41
Babel==2.8.0
bcrypt==3.1.7
botocore==1.15.41
cffi==1.14.0
click==7.1.1
colorama==0.4.3
cryptography==2.9
distlib==0.3.0
docutils==0.15.2
ecdsa==0.15
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.11.1
jmespath==0.9.5
jwt==1.0.0
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
parso==0.6.2
pluggy==0.13.1
psycopg2==2.8.5
psycopg2-binary==2.8.4
py==1.8.1
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
pydocstyle==5.0.2
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.6.0
python-editor==1.0.4
python-jose==3.1.0
python-jsonrpc-server==0.3.4
python-language-server==0.31.8
pytz==2019.3
PyYAML==5.3.1
rope==0.16.0
rsa==3.4.2
s3transfer==0.3.3
six==1.14.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.15
typed-ast==1.4.1
ujson==1.35
urllib3==1.25.9
virtualenv==20.0.10
wcwidth==0.1.9
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
yapf==0.29.0
zipp==3.1.0
